iz-ingest-status/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
regex==2024.9.11
mysql-connector-python==9.1.0
pydantic-settings==2.5.0
python-multipart==0.0.12
croniter==2.0.7
pytest==8.3.3
httpx==0.27.2
